+++ packages/web/src/components/ui/card/index.tsx
@@ -86,14 +86,6 @@ const CardDescription = ({ as: Component = 'p', className, ...props }: CardDescr
 
 CardDescription.displayName = 'Card.Description';
 
-const CardLegend = forwardRef<HTMLParagraphElement, HTMLAttributes<HTMLParagraphElement>>(
-  ({ className, ...props }, ref) => (
-    <p ref={ref} className={cx('text-sm italic text-muted', className)} {...props} />
-  )
-);
-
-CardLegend.displayName = 'Card.Legend';
-
 interface CardContentProps extends HTMLAttributes<HTMLDivElement> {}
 
 const CardContent = forwardRef<HTMLDivElement, CardContentProps>(({ className, ...props }, ref) => (
@@ -153,7 +145,6 @@ const Card = Object.assign(CardRoot, {
   Title: CardTitle,
   Subtitle: CardSubtitle,
   Description: CardDescription,
-  Legend: CardLegend,
   Content: CardContent,
   Footer: CardFooter,
   Badge: CardBadge